Renamed package "protocol" to "auth"
