+++ auth/authentication.go
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package protocol
+package auth
 
 import (
 	"encoding/base64"
