+++ streaming/proxy_test.go
@@ -19,7 +19,7 @@ package streaming
 import (
 	"encoding/hex"
 	"github.com/onitake/restreamer/configuration"
-	"github.com/onitake/restreamer/protocol"
+	"github.com/onitake/restreamer/auth"
 	"github.com/onitake/restreamer/util"
 	"net/http"
 	"net/url"
@@ -89,7 +89,7 @@ func testWithProxy(t *testing.T, proxy *Proxy) {
 func TestProxy(t *testing.T) {
 	l := &mockProxyLogger{t}
 
-	auth := protocol.NewAuthenticator(configuration.Authentication{}, nil)
+	auth := auth.NewAuthenticator(configuration.Authentication{}, nil)
 
 	direct, _ := NewProxy("file:///tmp/test.txt", 10, 0, auth)
 	logger = l
