+++ streaming/streamer.go
@@ -22,7 +22,7 @@ import (
 	"github.com/onitake/restreamer/api"
 	"github.com/onitake/restreamer/event"
 	"github.com/onitake/restreamer/mpegts"
-	"github.com/onitake/restreamer/protocol"
+	"github.com/onitake/restreamer/auth"
 	"github.com/onitake/restreamer/util"
 	"net/http"
 	"sync"
@@ -101,7 +101,7 @@ type Streamer struct {
 	// events is an event receiver
 	events event.Notifiable
 	// auth is an authentication verifier for client requests
-	auth protocol.Authenticator
+	auth auth.Authenticator
 }
 
 // ConnectionBroker represents a policy handler for new connections.
@@ -121,7 +121,7 @@ type ConnectionBroker interface {
 // qsize is the length of each connection's queue (in packets).
 // broker handles policy enforcement
 // stats is a statistics collector object.
-func NewStreamer(qsize uint, broker ConnectionBroker, auth protocol.Authenticator) *Streamer {
+func NewStreamer(qsize uint, broker ConnectionBroker, auth auth.Authenticator) *Streamer {
 	streamer := &Streamer{
 		broker:    broker,
 		queueSize: int(qsize),
@@ -332,7 +332,7 @@ func (streamer *Streamer) Stream(queue <-chan mpegts.Packet) error {
 // Satisfies the http.Handler interface, so it can be used in an HTTP server.
 func (streamer *Streamer) ServeHTTP(writer http.ResponseWriter, request *http.Request) {
 	// fail-fast: verify that this user can access this resource first
-	if !protocol.HandleHttpAuthentication(streamer.auth, request, writer) {
+	if !auth.HandleHttpAuthentication(streamer.auth, request, writer) {
 		return
 	}
